Fixed details of logic ans styles
